Jira id: HCLANG-61 - #comment refactored test method names
